+++ tests/run.lua
@@ -21,6 +21,7 @@ local specs = {
   "tests.spec.mode_spec",
   "tests.spec.system_spec",
   "tests.spec.lifecycle_spec",
+  "tests.spec.defaults_spec",
 }
 
 -- A spec that throws while loading used to abort this file, which meant